@@ -19,8 +19,8 @@ package com.robinhood.spark;
 import android.database.DataSetObservable;
 import android.database.DataSetObserver;
 import android.graphics.RectF;
-import android.support.annotation.NonNull;
-import android.support.annotation.VisibleForTesting;
+import androidx.annotation.NonNull;
+import androidx.annotation.VisibleForTesting;
 
 /**
  * A simple adapter class - evenly distributes your points along the x axis, does not draw a base
